Refactored error handling

+++ backend/src/manager.rs
@@ -1,15 +1,5 @@
-use log::{error, info, warn};
-use serde::Serialize;
-use std::{
-    collections::{BTreeMap, HashSet},
-    error::Error,
-    sync::{Arc, Mutex},
-    thread::{self, JoinHandle},
-    time::Duration,
-};
-use tokio::runtime::Runtime;
-
 use crate::{
+    error::{Error, Result},
     kubernetes::{Configuration, Engine, Environment},
     metrics::Metrics,
     types::{
@@ -17,17 +7,15 @@ use crate::{
         Template, User, UserConfiguration, UserUpdateConfiguration,
     },
 };
-
-/*
-
-https://github.com/clux/kube-rs/blob/master/examples/event_watcher.rs
-https://github.com/kubesphere/kube-events
-https://github.com/heptiolabs/eventrouter
-https://github.com/opsgenie/kubernetes-event-exporter
-https://github.com/GoogleCloudPlatform/click-to-deploy/blob/master/k8s/prometheus/manifest/prometheus-statefulset.yaml
-https://github.com/prometheus-operator/kube-prometheus
-
-*/
+use log::{error, info, warn};
+use serde::Serialize;
+use std::{
+    collections::{BTreeMap, HashSet},
+    sync::{Arc, Mutex},
+    thread::{self, JoinHandle},
+    time::Duration,
+};
+use tokio::runtime::Runtime;
 
 fn running_sessions(sessions: Vec<&Session>) -> Vec<&Session> {
     sessions
@@ -54,8 +42,8 @@ pub struct Playground {
 impl Manager {
     const SLEEP_TIME: Duration = Duration::from_secs(60);
 
-    pub async fn new() -> Result<Self, Box<dyn Error>> {
-        let metrics = Metrics::new()?;
+    pub async fn new() -> Result<Self> {
+        let metrics = Metrics::new().map_err(|err| Error::Failure(err.into()))?;
         let engine = Engine::new().await?;
         // Go through all existing sessions and update the ingress
         match engine.clone().list_sessions().await {
@@ -89,74 +77,76 @@ impl Manager {
             thread::sleep(Manager::SLEEP_TIME);
 
             // Track some deployments metrics
-            let sessions_thread = self.clone().sessions.clone();
-            if let Ok(mut sessions2) = sessions_thread.lock() {
-                let sessions3 = &mut sessions2.clone();
-                for id in sessions3.iter() {
-                    match self.clone().get_session(id) {
-                        Ok(Some(session)) => {
-                            // Deployed sessions are removed from the set
-                            // Additionally the deployment time is tracked
-                            match session.pod.phase {
-                                Phase::Running | Phase::Failed => {
-                                    sessions2.remove(&session.user_id);
-                                    if let Some(duration) =
-                                        &session.pod.start_time.and_then(|p| p.elapsed().ok())
-                                    {
-                                        self.clone()
-                                            .metrics
-                                            .observe_deploy_duration(duration.as_secs_f64());
-                                    } else {
-                                        error!("Failed to compute this session lifetime");
+            if let Ok(runtime) = new_runtime() {
+                let sessions_thread = self.clone().sessions.clone();
+                if let Ok(mut sessions2) = sessions_thread.lock() {
+                    let sessions3 = &mut sessions2.clone();
+                    for id in sessions3.iter() {
+                        match runtime.block_on(self.engine.get_session(&session_id(id))) {
+                            Ok(Some(session)) => {
+                                // Deployed sessions are removed from the set
+                                // Additionally the deployment time is tracked
+                                match session.pod.phase {
+                                    Phase::Running | Phase::Failed => {
+                                        sessions2.remove(&session.user_id);
+                                        if let Some(duration) =
+                                            &session.pod.start_time.and_then(|p| p.elapsed().ok())
+                                        {
+                                            self.clone()
+                                                .metrics
+                                                .observe_deploy_duration(duration.as_secs_f64());
+                                        } else {
+                                            error!("Failed to compute this session lifetime");
+                                        }
                                     }
-                                }
-                                _ => {}
-                            }
-                        }
-                        Err(err) => {
-                            warn!("Failed to call get: {}", err);
-                            sessions2.remove(id);
-                        }
-                        Ok(None) => warn!("No matching pod: {}", id),
-                    }
-                }
-            } else {
-                error!("Failed to acquire sessions lock");
-            }
-
-            // Go through all Running pods and figure out if they have to be undeployed
-            match self.clone().list_sessions() {
-                Ok(sessions) => {
-                    for session in running_sessions(sessions.values().collect()) {
-                        if let Some(duration) =
-                            &session.pod.start_time.and_then(|p| p.elapsed().ok())
-                        {
-                            if duration > &session.duration {
-                                info!("Undeploying {}", session.user_id);
-
-                                match self.clone().delete_session(&session.user_id) {
-                                    Ok(()) => (),
-                                    Err(err) => {
-                                        warn!(
-                                            "Error while undeploying {}: {}",
-                                            session.user_id, err
-                                        )
-                                    }
+                                    _ => {}
                                 }
                             }
-                        } else {
-                            error!("Failed to compute this session lifetime");
+                            Err(err) => {
+                                warn!("Failed to call get: {}", err);
+                                sessions2.remove(id);
+                            }
+                            Ok(None) => warn!("No matching pod: {}", id),
                         }
                     }
+                } else {
+                    error!("Failed to acquire sessions lock");
                 }
-                Err(err) => error!("Failed to call list_all: {}", err),
+
+                // Go through all Running pods and figure out if they have to be undeployed
+                match runtime.block_on(self.engine.list_sessions()) {
+                    Ok(sessions) => {
+                        for session in running_sessions(sessions.values().collect()) {
+                            if let Some(duration) =
+                                &session.pod.start_time.and_then(|p| p.elapsed().ok())
+                            {
+                                if duration > &session.duration {
+                                    info!("Undeploying {}", session.user_id);
+
+                                    match runtime.block_on(self.engine.delete_session(&session_id(&session.user_id))) {
+                                        Ok(()) => (),
+                                        Err(err) => {
+                                            warn!(
+                                                "Error while undeploying {}: {}",
+                                                session.user_id, err
+                                            )
+                                        }
+                                    }
+                                }
+                            } else {
+                                error!("Failed to compute this session lifetime");
+                            }
+                        }
+                    }
+                    Err(err) => error!("Failed to call list_all: {}", err),
+                }
             }
         })
     }
 }
 
-fn new_runtime() -> Result<Runtime, String> {
-    Runtime::new().map_err(|err| format!("{}", err))
+fn new_runtime() -> Result<Runtime> {
+    Runtime::new().map_err(|err| Error::Failure(err.into()))
 }
 
 fn session_id(id: &str) -> String {
@@ -165,7 +155,7 @@ fn session_id(id: &str) -> String {
 }
 
 impl Manager {
-    pub fn get(self, user: LoggedUser) -> Result<Playground, String> {
+    pub fn get(self, user: LoggedUser) -> Result<Playground> {
         let templates = new_runtime()?.block_on(self.clone().engine.list_templates())?;
         Ok(Playground {
             templates,
@@ -175,70 +165,107 @@ impl Manager {
         })
     }
 
-    pub fn get_unlogged(self) -> Result<Playground, String> {
+    pub fn get_unlogged(&self) -> Result<Playground> {
         let templates = new_runtime()?.block_on(self.clone().engine.list_templates())?;
         Ok(Playground {
             templates,
             user: None,
-            env: self.engine.env,
-            configuration: self.engine.configuration,
+            env: self.clone().engine.env,
+            configuration: self.clone().engine.configuration,
         })
     }
 
     // Users
 
-    pub fn get_user(self, id: &str) -> Result<Option<User>, String> {
-        new_runtime()?.block_on(self.engine.get_user(&id))
+    pub fn get_user(&self, user: &LoggedUser, id: &str) -> Result<Option<User>> {
+        if user.has_admin_read_rights() {
+            new_runtime()?.block_on(self.engine.get_user(&id))
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
-    pub fn list_users(self) -> Result<BTreeMap<String, User>, String> {
-        new_runtime()?.block_on(self.engine.list_users())
+    pub fn list_users(&self, user: &LoggedUser) -> Result<BTreeMap<String, User>> {
+        if user.has_admin_read_rights() {
+            new_runtime()?.block_on(self.engine.list_users())
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
-    pub fn create_user(self, id: String, user: UserConfiguration) -> Result<(), String> {
-        new_runtime()?.block_on(self.engine.create_user(id, user))
+    pub fn create_user(self, user: &LoggedUser, id: String, conf: UserConfiguration) -> Result<()> {
+        if user.has_admin_edit_rights() {
+            new_runtime()?.block_on(self.engine.create_user(id, conf))
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
-    pub fn update_user(self, id: String, user: UserUpdateConfiguration) -> Result<(), String> {
-        new_runtime()?.block_on(self.engine.update_user(id, user))
+    pub fn update_user(
+        self,
+        user: LoggedUser,
+        id: String,
+        conf: UserUpdateConfiguration,
+    ) -> Result<()> {
+        if user.has_admin_edit_rights() {
+            new_runtime()?.block_on(self.engine.update_user(id, conf))
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
-    pub fn delete_user(self, id: String) -> Result<(), String> {
-        new_runtime()?.block_on(self.engine.delete_user(id))
+    pub fn delete_user(self, user: &LoggedUser, id: String) -> Result<()> {
+        if user.has_admin_edit_rights() {
+            new_runtime()?.block_on(self.engine.delete_user(id))
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
     // Sessions
 
-    pub fn get_session(&self, id: &str) -> Result<Option<Session>, String> {
+    pub fn get_session(&self, user: &LoggedUser, id: &str) -> Result<Option<Session>> {
+        if !user.has_admin_read_rights() {
+            return Err(Error::Unauthorized());
+        }
+
         new_runtime()?.block_on(self.engine.get_session(&session_id(id)))
     }
 
-    pub fn list_sessions(&self) -> Result<BTreeMap<String, Session>, String> {
-        new_runtime()?.block_on(self.engine.list_sessions())
+    pub fn list_sessions(&self, user: &LoggedUser) -> Result<BTreeMap<String, Session>> {
+        if user.has_admin_read_rights() {
+            new_runtime()?.block_on(self.engine.list_sessions())
+        } else {
+            Err(Error::Unauthorized())
+        }
     }
 
     pub fn create_session(
-        self,
-        id: &str,
+        &self,
         user: &LoggedUser,
+        id: &str,
         conf: SessionConfiguration,
-    ) -> Result<(), String> {
+    ) -> Result<()> {
+        if !user.has_admin_edit_rights() {
+            return Err(Error::Unauthorized());
+        }
+
         if conf.duration.is_some() {
             // Duration can only customized by users with proper rights
             if !user.can_customize_duration() {
-                return Err("Only admin can customize a session duration".to_string());
+                return Err(Error::Unauthorized());
             }
         }
         if conf.pool_affinity.is_some() {
             // Duration can only customized by users with proper rights
             if !user.can_customize_pool_affinity() {
-                return Err("Only admin can customize a session pool affinity".to_string());
+                return Err(Error::Unauthorized());
             }
         }
 
         let session_id = session_id(id);
-        if self.get_session(&session_id)?.is_some() {
-            return Err("A session is already running".to_string());
+        if self.get_session(user, &session_id)?.is_some() {
+            return Err(Error::Unauthorized());
         }
 
         let template = conf.clone().template;
@@ -264,21 +291,29 @@ impl Manager {
     }
 
     pub fn update_session(
-        self,
+        &self,
         id: &str,
         user: &LoggedUser,
         conf: SessionUpdateConfiguration,
-    ) -> Result<(), String> {
+    ) -> Result<()> {
+        if !user.has_admin_edit_rights() {
+            return Err(Error::Unauthorized());
+        }
+
         if conf.duration.is_some() {
             // Duration can only customized by users with proper rights
             if !user.can_customize_duration() {
-                return Err("Only admin can customize a session duration".to_string());
+                return Err(Error::Unauthorized());
             }
         }
         new_runtime()?.block_on(self.engine.update_session(&session_id(id), conf))
     }
 
-    pub fn delete_session(self, id: &str) -> Result<(), String> {
+    pub fn delete_session(&self, user: &LoggedUser, id: &str) -> Result<()> {
+        if !user.has_admin_edit_rights() {
+            return Err(Error::Unauthorized());
+        }
+
         let session_id = session_id(id);
         let result = new_runtime()?.block_on(self.engine.delete_session(&session_id));
         match &result {
@@ -300,11 +335,19 @@ impl Manager {
 
     // Pools
 
-    pub fn get_pool(self, id: &str) -> Result<Option<Pool>, String> {
-        new_runtime()?.block_on(self.engine.get_pool(&id))
+    pub fn get_pool(&self, user: &LoggedUser, pool_id: &str) -> Result<Option<Pool>> {
+        if !user.has_admin_read_rights() {
+            return Err(Error::Unauthorized());
+        }
+
+        new_runtime()?.block_on(self.engine.get_pool(&pool_id))
     }
 
-    pub fn list_pools(&self) -> Result<BTreeMap<String, Pool>, String> {
+    pub fn list_pools(&self, user: &LoggedUser) -> Result<BTreeMap<String, Pool>> {
+        if !user.has_admin_read_rights() {
+            return Err(Error::Unauthorized());
+        }
+
         new_runtime()?.block_on(self.clone().engine.list_pools())
     }
 }
